# Optimized Background Noise Augmentation for Large Background Files