**Proposed Algorithm:**

    Get duration of the background file (bg_file) in seconds.
    Sample a random value from the range [0, bg_file_seconds - event_file_seconds).
    Read the background file from sampled_value to sampled_value + event_file_seconds.

This approach ensures that:

    We only load a portion of the background file required for the augmentation.
    It maintains randomness in background selection while reducing memory overhead.
    It is adaptable to cases with varied sample rates and event/background file durations.

**Experiments and Results:**

I’ve tested this algorithm using:

    Event durations ranging from 1 to 9 seconds.
    Background durations ranging from 81 to 10,000 seconds.
    Sample rates: 16,000 Hz, 22,500 Hz, and 44,100 Hz.

This optimized approach significantly reduces memory usage while maintaining augmentation quality. I’ve attached the comparison plot showcasing the performance difference for your reference.


    Improves scalability by avoiding unnecessary memory consumption for large files.
    Enhances performance in real-time audio augmentation workflows.
    Can be integrated as a feature or an option in AddBackgroundNoise to provide more flexibility to users.

Please let me know your thoughts on this proposal and if any further details or clarifications are needed.

In the below figure First plot shows difference between memory usage over the test_cases normalized for by 1e6 and next graph is time taken comparison of old vs proposed.

![Labeled_final](https://github.com/user-attachments/assets/94867218-461f-47a8-9bc4-506de010cb78)