@iver56 2025-01-15
In this case I would prefer lazy duration caching over eager caching. The difference becomes quite noticeable when there is a large number of files. If you have half a million files, and it takes 1 ms to check the duration of each file, initializing the class would take 500 seconds.