@iver56 2025-01-16
Thanks for implementing that change

> I do have a question regarding the lookup mechanism for file time information. Currently, I am using a dictionary for this purpose, but its average-case time complexity for lookups is not guaranteed to be constant. I am exploring an alternative approach using an array of size len(sound_file_paths). With this method:
> 
> * Each file would be assigned an index (e.g., from 0 to len(sound_file_paths) - 1).
> * File paths and corresponding time information could then be accessed directly using the index, enabling efficient retrieval.

* dict lookups are O(1) on average for both string and integer keys
* Having integers as keys is faster than having strings as keys, due to faster hashing and comparison. And it uses less memory.
* Accessing a value in an array/list is also O(1), but in practice it is faster than a dict lookup
* a numpy array requires less memory than a python-native list of floats

Here's a rough comparison of the memory usage in the three different alternatives, given that there are half a million items:
List of floats: ~16 MB
Dictionary (int keys, float values): ~47 MB
NumPy array (float32): ~2 MB

If you feel like optimizing it with your array idea, here's my green light: 🟢