@worldveil 2025-04-20
> > Additionally, I was wondering if there’s any provision in the current system to prioritize sampling certain files more frequently than others—for instance, based on importance, weight, or any custom-defined priority. If such functionality does not currently exist, is there a plan to introduce it in the future?
> 
> I don't have any immediate plans for adding that feature, but you're welcome to add an issue for it

@PratikKulkar see my PR [here](https://github.com/iver56/audiomentations/pull/387) for this feature!