@PratikKulkar 2025-08-05
> Hey @PratikKulkar thanks for the work you've done I missed your PR and ended up implementing the same thing 😅. Maybe we can use this PR to remove the `lru_cache `from this augmentation as it won't be working properly in this set up anyway. Not sure how @iver56 sees this as it will be a breaking change. Also we can apply the same idea to `ApplyImpulseResponse` but I guess it's better to open another PR for the sake of clarity.

Hey @JorisCos, no worries at all — these things happen 😅
Glad we were thinking along the same lines! I agree, we can definitely repurpose this PR to remove the lru_cache from the augmentation — especially since it doesn’t behave well with this partial loading setup and could lead to unexpected memory issues.

As for ApplyImpulseResponse, yeah, makes sense to handle that in a separate PR for clarity.

Let me know what @iver56 thinks about the breaking change here, and I can update the PR accordingly.