@iver56 2025-08-05
Here's my recommendation: Set the default value of the `lru_cache_size` to `None`. If a value other than `None` gets passed, raise a `TypeError` exception explaining that that the LRU cache functionality has been removed, and it is no longer valid to set the `lru_cache_size` param.

You can remove the functools.lru_cache stuff accordingly. As audiomentations is still in alpha, it's technically okay to do breaking changes, but it's still nice to have them be well explained (clear how to make it compatible with new version) and have a good reason for doing it.

Thanks for the help, @PratikKulkar